Add changes to entity and migration files

+++ src/database/mysql/migrations/1696322819786-create-pairs-liquidity-changes.ts
@@ -6,13 +6,13 @@ export class createPairsLiquidityChanges1696322819786
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(
       `CREATE TABLE IF NOT EXISTS pairs_liquidity_changes ( \
-            id bigint unsigned NOT NULL AUTO_INCREMENT, \
+            id int NOT NULL AUTO_INCREMENT, \
             first_asset_id varchar(128) NOT NULL, \
             second_asset_id varchar(128) NOT NULL, \
             first_asset_amount varchar(256) NOT NULL, \
             second_asset_amount varchar(256) NOT NULL, \
             type varchar(16) NOT NULL, \
-            timestamp int NOT NULL, \
+            timestamp bigint NOT NULL, \
             PRIMARY KEY (id) \
         )`,
     );
